Add generate infra dynamic inventory playbook/role
Some clouds may not have metadata enabled, or even have it but is
not reliable.
Force the user to define in infra_config.yml the services that make
up the OpenStack Infra cloud. This will be in a dict named 'services',
which will contain an entry for each OpenStack Infra service
(puppetmaster, zuul, nodepool, etc).

Change-Id: I9e3fe34684cd31bf1750ad8cf72e7e9090bfe087